tagging version 1.3.4
git-svn-id: https://plugins.svn.wordpress.org/geodirectory/trunk@1064297 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ geodirectory-assets/js/geodirectory-plupload.js
@@ -69,7 +69,7 @@ jQuery(document).ready(function($) {
 			jQuery('#'+imgId+'upload-error').removeClass('upload-error');
 			
 			
-			if(limitImg && $this.hasClass("plupload-upload-uic-multiple")){
+			if(limitImg && $this.hasClass("plupload-upload-uic-multiple") && jQuery("#"+imgId+"image_limit").val()){
 				
 				if(totalImg==limitImg && parseInt(limitImg) > 0){
 				while(up.files.length > 0) {up.removeFile(up.files[0]);} // remove images
